+++ src/Controller/CaseController.php
@@ -128,10 +128,6 @@ class CaseController extends AbstractController
 
         $filterBuilderUpdater->addFilterConditions($filterForm, $filterBuilder);
 
-        // Add sortable fields depending on case type.
-        $filterBuilder->leftJoin('c.complaintCategory', 'complaintCategory');
-        $filterBuilder->addSelect('partial complaintCategory.{id,name}');
-
         // Only get agendas under active municipality
         $filterBuilder->andWhere('c.municipality = :municipality')
             ->setParameter('municipality', $activeMunicipality->getId()->toBinary())